dpkgbuild + warning fixed

+++ src/MPFDParser/Parser.cc
@@ -45,7 +45,7 @@ void MPFD::Parser::SetContentType(const std::string type) {
     }
 
 
-    int bp = type.find("boundary=");
+    size_t bp = type.find("boundary=");
 
     if (bp == std::string::npos) {
         throw MPFD::Exception(std::string("Cannot find boundary in Content-type: \"") + type + std::string("\""));
@@ -169,11 +169,11 @@ void MPFD::Parser::_ParseHeaders(std::string headers) {
     }
 
     // Find name
-    long name_pos = headers.find("name=\"");
+    size_t name_pos = headers.find("name=\"");
     if (name_pos == std::string::npos) {
         throw Exception(std::string("Accepted headers of field does not contain \"name=\".\nThe headers are: \"") + headers + std::string("\""));
     } else {
-        long name_end_pos = headers.find("\"", name_pos + 6);
+        size_t name_end_pos = headers.find("\"", name_pos + 6);
         if (name_end_pos == std::string::npos) {
             throw Exception(std::string("Cannot find closing quote of \"name=\" attribute.\nThe headers are: \"") + headers + std::string("\""));
         } else {
@@ -183,7 +183,7 @@ void MPFD::Parser::_ParseHeaders(std::string headers) {
 
 
         // find filename if exists
-        long filename_pos = headers.find("filename=\"");
+        size_t filename_pos = headers.find("filename=\"");
         if (filename_pos == std::string::npos) {
             Fields[ProcessingFieldName]->SetType(Field::TextType);
         } else {
@@ -191,7 +191,7 @@ void MPFD::Parser::_ParseHeaders(std::string headers) {
             Fields[ProcessingFieldName]->SetTempDir(TempDirForFileUpload);
             Fields[ProcessingFieldName]->SetUploadedFilesStorage(WhereToStoreUploadedFiles);
 
-            long filename_end_pos = headers.find("\"", filename_pos + 10);
+            size_t filename_end_pos = headers.find("\"", filename_pos + 10);
             if (filename_end_pos == std::string::npos) {
                 throw Exception(std::string("Cannot find closing quote of \"filename=\" attribute.\nThe headers are: \"") + headers + std::string("\""));
             } else {
@@ -200,10 +200,10 @@ void MPFD::Parser::_ParseHeaders(std::string headers) {
             }
 
             // find Content-Type if exists
-            long content_type_pos = headers.find("Content-Type: ");
+            size_t content_type_pos = headers.find("Content-Type: ");
             if (content_type_pos != std::string::npos) {
-                long content_type_end_pos = 0;
-                for (int i = content_type_pos + 14; (i < headers.length()) && (!content_type_end_pos); i++) {
+                size_t content_type_end_pos = 0;
+                for (size_t i = content_type_pos + 14; (i < headers.length()) && (!content_type_end_pos); i++) {
                     if ((headers[i] == ' ') || (headers[i] == 10) || (headers[i] == 13)) {
                         content_type_end_pos = i - 1;
                     }
